requirements.txt

fastapi==0.110.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
sqlalchemy==2.0.31
pgvector==0.2.5
psycopg[binary]==3.2.1
requests==2.32.3
python-dotenv==1.0.1
sentence-transformers==2.7.0
numpy==1.26.4
rapidfuzz==3.9.6
dateparser==1.2.0
jinja2==3.1.4